送信責任をPageに移行
